requirements.txt
numpy
opencv-python
matplotlib
python-telegram-bot